+++ packages/edition-slices/.jestlint
@@ -1,5 +1,5 @@
 {
   "maxAttributeStringLength": 150,
-  "maxFileSize": 56393,
+  "maxFileSize": 60500,
   "maxLines": 650
 }
