+++ packages/edition-slices/.prettierignore
@@ -12,3 +12,4 @@ LICENSE
 yarn.lock
 *.log
 rnw.js
+assets
