+++ packages/edition-slices/assets/leaders-masthead.png
