+++ packages/edition-slices/src/slices/leaders/styles.js
@@ -1,0 +1,24 @@
+import { fonts, spacing } from "@times-components/styleguide";
+
+const styles = {
+  imageStyle: {
+    height: 51,
+    margin: "auto",
+    width: 283
+  },
+  leadText: {
+    color: "#850029",
+    fontFamily: fonts.bodyRegular,
+    fontSize: 16,
+    lineHeight: spacing(3)
+  },
+  leadTextContainer: {
+    paddingBottom: spacing(1),
+    paddingTop: spacing(2)
+  },
+  text: {
+    textAlign: "center"
+  }
+};
+
+export default styles;
