+++ packages/slice-layout/__tests__/leaders.base.js
@@ -10,6 +10,7 @@ export default renderComponent => {
       test() {
         const output = renderComponent(
           <Leaders
+            renderHead={() => createItem("leader-4")}
             renderLeader1={() => createItem("leader-1")}
             renderLeader2={() => createItem("leader-2")}
             renderLeader3={() => createItem("leader-3")}
