+++ packages/slice-layout/src/templates/leaders/index.js
@@ -1,21 +1,18 @@
 import React from "react";
-import { View, Text, Image } from "react-native";
+import { View } from "react-native";
 import styles from "./styles";
 import propTypes from "./proptypes";
 
 const keySeperator = () => <View style={styles.keySeperator} />;
 
-const Leaders = ({ renderLeader1, renderLeader2, renderLeader3 }) => (
+const Leaders = ({
+  renderLeader1,
+  renderLeader2,
+  renderLeader3,
+  renderHead
+}) => (
   <View style={styles.container}>
-    <Image
-      resizeMode="contain"
-      // eslint-disable-next-line global-require
-      source={require("../../../assets/leaders-masthead.png")}
-      style={styles.imageStyle}
-    />
-    <View style={styles.leadTextContainer}>
-      <Text style={[styles.leadText, styles.text]}> Leading Articles </Text>
-    </View>
+    {renderHead()}
     {renderLeader1()}
     {keySeperator()}
     {renderLeader2()}
