+++ packages/slice-layout/src/templates/leaders/proptypes.js
@@ -1,6 +1,7 @@
 import PropTypes from "prop-types";
 
 const propTypes = {
+  renderHead: PropTypes.func.isRequired,
   renderLeader1: PropTypes.func.isRequired,
   renderLeader2: PropTypes.func.isRequired,
   renderLeader3: PropTypes.func.isRequired
