+++ packages/slice-layout/src/templates/leaders/styles.js
@@ -1,4 +1,4 @@
-import { fonts, spacing } from "@times-components/styleguide";
+import { spacing } from "@times-components/styleguide";
 
 const styles = {
   container: {
@@ -6,29 +6,11 @@ const styles = {
     paddingHorizontal: spacing(8),
     paddingTop: spacing(4)
   },
-  imageStyle: {
-    height: 51,
-    margin: "auto",
-    width: 283
-  },
   keySeperator: {
     borderBottomWidth: 1,
     borderColor: "#dcdcdc",
     borderStyle: "solid",
     marginHorizontal: spacing(2)
-  },
-  leadText: {
-    color: "#850029",
-    fontFamily: fonts.bodyRegular,
-    fontSize: 16,
-    lineHeight: spacing(3)
-  },
-  leadTextContainer: {
-    paddingBottom: spacing(1),
-    paddingTop: spacing(2)
-  },
-  text: {
-    textAlign: "center"
   }
 };
